committing this because I can't remember what I did here

+++ src/Constants.hh
@@ -5,6 +5,6 @@ namespace Constants {
     const float screenHeight = 650.f;
     const float screenWidth = 900.f;
 
-    const int debugScreenHeight = screenHeight + 200.f;
-    const int debugScreenWidth = screenWidth + 600.f;
+    const int debugScreenHeight = static_cast<const int>(screenHeight + 200.f);
+    const int debugScreenWidth = static_cast<const int>(screenWidth + 600.f);
 }
